DatabaseController split off from SystemController DatabaseController used in MetaDataController and SystemController
